blogpost_api/requirements.txt
asgiref==3.2.3
Brotli==1.0.7
certifi==2019.11.28
chardet==3.0.4
checkexpect==1.0.19
click==6.7
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
docopt==0.6.2
feedparser==5.2.1
gitsome==0.8.4
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
ply==3.11
prompt-toolkit==2.0.10
psycopg2==2.8.4
psycopg2-binary==2.8.4
Pygments==2.5.2
python-slugify==4.0.0
pytz==2016.10
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
uritemplate==3.0.1
uritemplate.py==3.0.2
urllib3==1.25.8
wcwidth==0.1.8
whitenoise==5.0.1
